requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
APScheduler==3.7.0
async-timeout==3.0.1
attrs==21.2.0
black==21.7b0
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.5
chat-exporter==1.7
click==8.0.1
debugpy==1.4.3
decorator==5.1.0
discord-ext-menus==1.1
discord.py==1.7.3
DiscordUtils==1.3.4
dnspython==2.1.0
emoji==1.4.2
enhanced-dpy==1.7.3.7.post1
grapheme==0.6.0
idna==3.2
multidict==4.7.6
mypy-extensions==0.4.3
pathspec==0.9.0
python-dotenv==0.19.0
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
tomli==1.2.1
tzlocal==2.1
urllib3==1.26.6
yarl==1.5.1
beautifulsoup4==4.10.0
ebaysdk==2.2.0
html5lib==1.1
lxml==4.6.3
emoji-country-flag==1.3.0
mongoengine==0.23.1